Updates `Board` component to use Vitest and new project structure
